Mark used kernel pages via linker start/end addresses
Lookup via page tables causes problems when kernel is mapped via large
pages (e.g. when using GRUB2) because the whole large page will be
marked as used, even when the kernel only occupies a small part of
it. -> No space available for heap because everything is marked as in
use by the kernel