Fix use-after-free in KMM
When enabling the DYNAMIC_KMM, a call to the freeSegment method
possibly removes the page, where the KMM saved the segment.
Any further access may lead to a page fault and subsequently to
a kernel panic.

For example, allocating two buffers (p1and p2) of at least the
size of PAGE_SIZE allocates at least two pages (assuming no
memory on the heap is free).
Freeing p1 does nothing because memory after p1 is not free.
Then freeing p2 leads to a call to ksbrk that unmaps the page of
p2, including its MallocSegment, and then the page of p1.
Further accesses to the MallocSegment of p2 fail and lead to a
panic.

The ksbrk method sets the kernel_break_, that a check can use to
see if heap-memory can be accessed safely.