Generate `compile_commands.json` by default
The `clangd` language server uses the `compile_commands.json` to find
the libraries and their respective path for showing lsp-features like
going to definition etc. The generation of this file can also be
triggered by configuring cmake like
`cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=1` but since there is a
`setup_cmake.sh` which wraps this it is more convenient to add it
directly to the `CMakeLists.txt` so that whatever way you configure
cmake it will always be there.